+++ src/controlador/PrestadorData.java
@@ -41,7 +41,7 @@ public class PrestadorData {
             statement.setString(2, prestador.getApellido());
             statement.setInt(3, prestador.getDni());
             statement.setString(4, prestador.getDomicilio());
-            statement.setInt(5, prestador.getTelefono());
+            statement.setLong(5, prestador.getTelefono());
             statement.setBoolean(6, prestador.isActivo());
             statement.setInt(7, prestador.getEspecialidad().getIdEspecialidad());
             statement.executeUpdate();
@@ -94,7 +94,7 @@ public class PrestadorData {
             statement.setString(2, prestador.getApellido());
             statement.setInt(3, prestador.getDni());
             statement.setString(4, prestador.getDomicilio());
-            statement.setInt(5, prestador.getTelefono());
+            statement.setLong(5, prestador.getTelefono());
             statement.setBoolean(6, prestador.isActivo());
             statement.setInt(7, prestador.getIdPrestador());
              int updateExitoso = statement.executeUpdate();
@@ -130,7 +130,7 @@ public class PrestadorData {
             prestador.setNombre(rs.getString("nombre"));
             prestador.setDni(rs.getInt("dni"));
             prestador.setDomicilio(rs.getString("domicilio"));
-            prestador.setTelefono(rs.getInt("telefono"));
+            prestador.setTelefono(rs.getLong("telefono"));
             prestador.setActivo(true);
             }
              else{
@@ -169,7 +169,7 @@ public class PrestadorData {
                prestador.setApellido(result.getString("apellido"));
                prestador.setDni(result.getInt("dni"));
                prestador.setDomicilio(result.getString("domicilio"));
-               prestador.setTelefono(result.getInt("telefono"));
+               prestador.setTelefono(result.getLong("telefono"));
                prestador.setActivo(result.getBoolean("activo"));
 
             } else {
